fix: Make new PageUpdate component look fresh
